+++ component/otelcol/config_tls.go
@@ -64,17 +64,6 @@ type TLSSetting struct {
 	ReloadInterval time.Duration     `river:"reload_interval,attr,optional"`
 }
 
-// UnmarshalRiver implements river.Unmarshaler and reports whether the
-// unmarshaled TLSConfig is valid.
-func (t *TLSSetting) UnmarshalRiver(f func(interface{}) error) error {
-	type tlsSetting TLSSetting
-	if err := f((*tlsSetting)(t)); err != nil {
-		return err
-	}
-
-	return t.Validate()
-}
-
 func (args *TLSSetting) Convert() *otelconfigtls.TLSSetting {
 	if args == nil {
 		return nil
@@ -93,7 +82,7 @@ func (args *TLSSetting) Convert() *otelconfigtls.TLSSetting {
 	}
 }
 
-// Validate reports whether t is valid.
+// Validate implements river.Validator.
 func (t *TLSSetting) Validate() error {
 	if len(t.CA) > 0 && len(t.CAFile) > 0 {
 		return fmt.Errorf("at most one of ca_pem and ca_file must be configured")