+++ pkg/flow/internal/controller/config_tracing.go
@@ -20,6 +20,8 @@ type TracingConfigNode struct {
 	eval  *vm.Evaluator
 }
 
+var _ BlockNode = (*TracingConfigNode)(nil)
+
 // NewTracingConfigNode creates a new TracingConfigNode from an initial ast.BlockStmt.
 // The underlying config isn't applied until Evaluate is called.
 func NewTracingConfigNode(block *ast.BlockStmt, globals ComponentGlobals) *TracingConfigNode {