+++ pkg/operator/hierarchy/hierarchy_test.go
@@ -125,7 +125,7 @@ func TestNotifier(t *testing.T) {
 			}
 
 			e := notifier.EventHandler()
-			e.Create(event.CreateEvent{Object: testPod}, q)
+			e.Create(ctx, event.CreateEvent{Object: testPod}, q)
 			if tc.expectEnqueue {
 				require.Equal(t, 1, q.Len(), "expected change enqueue")
 			} else {